feat: 기본 단일 toast 구조 생성
