--- packages/ui/src/toast/close.ts
@@ -1,3 +1,0 @@
-const autoClose = (time: number, id: string) => {
-  setTimeout(() => {}, time);
-};
